tools: move more support code to lib
