+++ tools/lib/entities_parser.py
