+++ tools/lib/map_parser.py
